added AI and damage on collision with AI

+++ src/components/entityComponents/HealthComponent.java
@@ -8,9 +8,10 @@ import components.IComponent;
  *
  */
 public class HealthComponent implements IComponent {
-	private int health;
+	
+	private double health;
 
-	public HealthComponent(int h) {
+	public HealthComponent(double h) {
 		health = h;
 	}
 	
@@ -23,11 +24,11 @@ public class HealthComponent implements IComponent {
 		return ComponentType.Health;
 	}
 
-	public int getHealth() {
+	public double getHealth() {
 		return health;
 	}
 
-	public void setHealth(int h) {
+	public void setHealth(double h) {
 		health = h;
 	}
